add auth to the project
