+++ app/layout.tsx
@@ -5,6 +5,7 @@ import { Navigation } from '@/components'
 import { Theme } from '@radix-ui/themes'
 import { Toaster } from 'react-hot-toast'
 import clsx from 'clsx'
+import { Providers } from '@/context'
 
 const inter = Inter({ subsets: ['latin'] })
 
@@ -21,11 +22,13 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body className={clsx(inter.className, 'bg-zinc-50')}>
-        <Toaster position="bottom-center" reverseOrder={false} />
-        <Theme>
-          <Navigation />
-          {children}
-        </Theme>
+        <Providers>
+          <Toaster position="bottom-center" reverseOrder={false} />
+          <Theme>
+            <Navigation />
+            {children}
+          </Theme>
+        </Providers>
       </body>
     </html>
   )
