+++ components/index.ts
@@ -1,3 +1,2 @@
 export * from './Navigation'
 export * from './Doodles'
-export * from './SocialsLogin'
